+++ lib/main.h
@@ -15,8 +15,9 @@ struct game {
 		SDL_Renderer *renderer;
 	} screen;
 	SDL_Surface **sprites;
-	SDL_Texture **sprite_textures;
 	SDL_Surface **font;
+	SDL_Texture **sprite_textures;
+	SDL_Texture *map_texture;
 	char cursor;
 	SDL_bool inventory;
 	SDL_bool fullscreen;