+++ lib/play.h
@@ -18,9 +18,10 @@ struct player {
 	unsigned char quantity[MAX_INV];
 };
 
-extern void	player_init(struct player *cur_player);
+extern void	player_init(struct worldmap *map, struct player *cur_player);
 extern void	player_quit(struct player *cur_player);
-extern void	move_player(struct worldmap *map, struct player *cur_player, int x, int y);
+extern void	move_player(struct game *cur_game, struct worldmap *map, struct player *cur_player, int x, int y);
 extern void	toggle_inv(struct game *cur_game);
+extern void	random_start(struct worldmap *map, struct player *cur_player);
 
 #endif